% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootstrapResult.R
\name{BootstrapResult}
\alias{BootstrapResult}
\title{Bootstrap Result}
\description{
Use this class to measure the performance of a survival learner
on a test dataset using bootstrapping.
See example.
}
\examples{
library(mlr3proba)
library(mlr3pipelines)
library(tibble)
library(dplyr, warn.conflicts = FALSE)

# Lung task - impute missing values
task = tsk('lung')
pre = po('encode', method = 'treatment') \%>>\%
      po('imputelearner', lrn('regr.rpart'))
task = pre$train(task)[[1]]
# partition to train and test sets
part = partition(task)

# train model on the train set
cox = lrn('surv.coxph', id = 'cox')
cox$train(task, row_ids = part$train)

# bootstrap the test set 10 times and measure performance
# using all available test metrics from `bench_msrs()`
# can be parallelized by increasing `test_workers`
bres = BootstrapResult$new(test_nrsmps = 10)
bres$calculate(task = task, learner = cox, part = part)

# performance on the test set
bres$score

# bootstrapped performance scores
bres$scores

# Median summary score per metric
bres$score_median()

# Percentile Confidence Intervals
bres$percent_ci()
bres$percent_ci(conf = 0.9)

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{task_id}}{Task id}

\item{\code{lrn_id}}{Learner id}

\item{\code{test_measure_ids}}{Measure ids}

\item{\code{test_nrsmps}}{Number of bootstrapped resamplings}

\item{\code{test_workers}}{Number of workers for bootstrap parallelization}

\item{\code{score}}{Performance on the original test set}

\item{\code{scores}}{Performance on bootstrapped test sets}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BootstrapResult-new}{\code{BootstrapResult$new()}}
\item \href{#method-BootstrapResult-calculate}{\code{BootstrapResult$calculate()}}
\item \href{#method-BootstrapResult-percent_ci}{\code{BootstrapResult$percent_ci()}}
\item \href{#method-BootstrapResult-score_median}{\code{BootstrapResult$score_median()}}
\item \href{#method-BootstrapResult-clone}{\code{BootstrapResult$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BootstrapResult-new"></a>}}
\if{latex}{\out{\hypertarget{method-BootstrapResult-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BootstrapResult$new(
  test_measure_ids = bench_msrs()$id,
  test_nrsmps = 1000,
  test_workers = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{test_measure_ids}}{Measure ids.
Default: use all available measures from \code{\link[=bench_msrs]{bench_msrs()}} function.}

\item{\code{test_nrsmps}}{Number of bootstrapped resamplings. Default: 1000.}

\item{\code{test_workers}}{Number of workers for bootstrap parallelization.
Default is 1 (sequential).
We advise careful consideration of how many workers to use.
Several factor might influence this, such as the total number of CPUs
and memory available, the size of the dataset, the learner's memory
footprint, the learner itself (e.g. \code{xgboost} might have problems), etc.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BootstrapResult-calculate"></a>}}
\if{latex}{\out{\hypertarget{method-BootstrapResult-calculate}{}}}
\subsection{Method \code{calculate()}}{
Calculate bootstrap performance scores on the test set of
the given survival task.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BootstrapResult$calculate(task, learner, part, quiet = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{A \link[mlr3proba:TaskSurv]{survival task}}

\item{\code{learner}}{A \link[mlr3proba:LearnerSurv]{learner}, \strong{trained} on the
\code{part$train} set of the given task.}

\item{\code{part}}{A \link[mlr3:partition]{partition} of the given task to train
and test sets.}

\item{\code{quiet}}{Default: TRUE (\strong{don't} report time elapsed)}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Parallelization: we check if the system allows for
\link[future:multicore]{multicore}
parallelization backend strategy (Linux systems mostly) and use that,
otherwise we use \link[future:multisession]{multisession}.
}

\subsection{Returns}{
(invisibly) a \link[tibble:tibble]{tibble} (\code{scores}) with columns the
metrics used and rows the performance scores measured on the different
bootstrapped resamplings.
Also the performance per metric on the original test set is recorded in
the variable \code{score}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BootstrapResult-percent_ci"></a>}}
\if{latex}{\out{\hypertarget{method-BootstrapResult-percent_ci}{}}}
\subsection{Method \code{percent_ci()}}{
Percentile confidence intervals per measure
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BootstrapResult$percent_ci(conf = 0.95)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{conf}}{Confidence level. Default: 0.95 (i.e. the 2.5th and 97.5th
percentile points are used to define the confidence interval range)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a \link[tibble:tibble]{tibble} with confidence intervals for the
performance score (per measure used).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BootstrapResult-score_median"></a>}}
\if{latex}{\out{\hypertarget{method-BootstrapResult-score_median}{}}}
\subsection{Method \code{score_median()}}{
Median summary score for the bootstrapped scores
(one per measure)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BootstrapResult$score_median()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a \link[tibble:tibble]{tibble row} with the median value of a score
across all bootstrap resamplings.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BootstrapResult-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BootstrapResult-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BootstrapResult$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
