% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fit_blme_model_cmp}
\alias{fit_blme_model_cmp}
\title{Fit Bayesian LME models to compare multiple models}
\usage{
fit_blme_model_cmp(
  df,
  nrsmps_perc = 1,
  n_chains = 4,
  n_iters = 5000,
  n_cores = 4,
  seed = 42,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{df}{a \code{data.frame}/\code{tibble} object with the benchmarking results.
Columns names must include: \code{task_id}, \code{lrn_id}, \code{rsmp_id}, \code{measure} and
\code{value}.
For example, see the output of \link{reshape_mob_res}.}

\item{nrsmps_perc}{Percentage of resampling ids to keep for faster
bayesian model fit (number between 0 and 1).
Default is 1, so we keep all of the resampling ids.}

\item{n_chains}{Number of Markon chains for MCMC. Default = 4}

\item{n_iters}{Number of iterations per chain. Default = 5000}

\item{n_cores}{Number of cores for MCMC parallelization. Default = 4}

\item{seed}{seed number for MCMC}

\item{verbose}{whether to print some informative messages. Default = TRUE.}

\item{...}{other arguments passed on to \link[rstanarm:stan_glmer]{rstanarm::stan_glmer}}
}
\value{
A list of \link[rstanarm:stanreg-objects]{stanreg-objects}, one per
performance \code{measure} in the benchmarking data.
}
\description{
Using a general format of benchmarking data, this function fits the following
Bayesian linear-mixed effects (LME) model using \link[rstanarm:stan_glmer]{stan_glmer}:

\deqn{value \sim -1 + lrn\_id + (1 | task\_id/rsmp\_id)}

where:
\itemize{
\item \code{lrn_id} is the model used
\item \code{task_id} is the dataset the model was used on
\item \code{rsmp_id} is the specific resampling of the dataset
\item \code{value} is the output value for a particular performance \code{measure}
}

One LME model is fit per \code{measure}, possibly in parallel using
\link[future.apply:future_lapply]{future_lapply}.
The aim is to get one posterior distribution per \code{model}, for each
performance \code{measure} in the benchmarking data, while accounting for
possible inter-dependencies between the different datasets as well as
the nested within-resample correlation that might exist.

See example in \link{MOBenchmark}.
}
