% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{mlr_measures_surv.oob_error}
\alias{mlr_measures_surv.oob_error}
\alias{MeasureSurvOOBError}
\title{Out-of-bag (OOB) Survival Error Measure}
\description{
Returns the out-of-bag error of a \link[mlr3proba:LearnerSurv]{LearnerSurv} if it
supports it (e.g. a Random Forest Survival learner that has the property
\code{'oob_error'}).
The \code{oob_error} is usually (1 - C-index) but it depends on the learner.
Returns \code{NA} for unsupported learners.
}
\section{Dictionary}{

This \link[mlr3:Measure]{Measure} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary}
\link[mlr3:mlr_measures]{mlr_measures} or with the associated sugar function \link[mlr3:mlr_sugar]{msr()}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{MeasureSurvOOBError$new()
mlr_measures$get('surv.oob_error')
msr('surv.oob_error')
}\if{html}{\out{</div>}}
}

\examples{
library(mlr3proba)
library(mlr3misc)
library(mlr3extralearners)

l = lrn('surv.ranger')
t = tsk('rats')
l$train(t)
p = l$predict(t)
m = msr('surv.oob_error')
p$score(m, learner = l)

}
\section{Super class}{
\code{\link[mlr3:Measure]{mlr3::Measure}} -> \code{MeasureSurvOOBError}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MeasureSurvOOBError-new}{\code{MeasureSurvOOBError$new()}}
\item \href{#method-MeasureSurvOOBError-clone}{\code{MeasureSurvOOBError$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="aggregate"><a href='../../mlr3/html/Measure.html#method-Measure-aggregate'><code>mlr3::Measure$aggregate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="format"><a href='../../mlr3/html/Measure.html#method-Measure-format'><code>mlr3::Measure$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="help"><a href='../../mlr3/html/Measure.html#method-Measure-help'><code>mlr3::Measure$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="print"><a href='../../mlr3/html/Measure.html#method-Measure-print'><code>mlr3::Measure$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="score"><a href='../../mlr3/html/Measure.html#method-Measure-score'><code>mlr3::Measure$score()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasureSurvOOBError-new"></a>}}
\if{latex}{\out{\hypertarget{method-MeasureSurvOOBError-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureSurvOOBError$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasureSurvOOBError-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MeasureSurvOOBError-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureSurvOOBError$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
