% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{PipeOpRemoveNAs}
\alias{PipeOpRemoveNAs}
\title{PipeOpRemoveNAs}
\description{
Removes features from a Task for which the proportion
of \code{NA} values are above a certain \code{cutoff} threshold.
}
\section{Parameters}{

\itemize{
\item \code{cutoff} :: \code{numeric(1)}\cr
Features with more than \code{cutoff} percentage of NAs will be removed.
Default value: 0.2.
}
}

\section{Initialization}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{PipeOpRemoveNAs$new()
pona = po('removenas')
}\if{html}{\out{</div>}}
}

\examples{
library(mlr3proba)
library(mlr3pipelines)

task = tsk('lung')
task
pona = PipeOpRemoveNAs$new()
pona$train(list(task))[[1L]] # meal.cal removed

pona$param_set$values$cutoff = 0 # remove every feature with at least 1 NA
pona$train(list(task))[[1L]]

# another way to do the same
pona = po('removenas', cutoff = 0.1)
pona$train(list(task))[[1L]] # meal.cal removed

}
\section{Super classes}{
\code{\link[mlr3pipelines:PipeOp]{mlr3pipelines::PipeOp}} -> \code{\link[mlr3pipelines:PipeOpTaskPreproc]{mlr3pipelines::PipeOpTaskPreproc}} -> \code{PipeOpRemoveNAs}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PipeOpRemoveNAs-new}{\code{PipeOpRemoveNAs$new()}}
\item \href{#method-PipeOpRemoveNAs-ncolsNA}{\code{PipeOpRemoveNAs$ncolsNA()}}
\item \href{#method-PipeOpRemoveNAs-clone}{\code{PipeOpRemoveNAs$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="help"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-help'><code>mlr3pipelines::PipeOp$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="predict"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-predict'><code>mlr3pipelines::PipeOp$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="print"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-print'><code>mlr3pipelines::PipeOp$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="train"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-train'><code>mlr3pipelines::PipeOp$train()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipeOpRemoveNAs-new"></a>}}
\if{latex}{\out{\hypertarget{method-PipeOpRemoveNAs-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipeOpRemoveNAs$new(id = "remove_na", param_vals = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character(1)})\cr
Identifier of the resulting  object.}

\item{\code{param_vals}}{(\code{list()})\cr
List of hyperparameter settings, overwriting the hyperparameter settings that would
otherwise be set during construction.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipeOpRemoveNAs-ncolsNA"></a>}}
\if{latex}{\out{\hypertarget{method-PipeOpRemoveNAs-ncolsNA}{}}}
\subsection{Method \code{ncolsNA()}}{
How many features (columns) have NAs (i.e. at least one)?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipeOpRemoveNAs$ncolsNA(task)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{\link[mlr3:Task]{Task}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipeOpRemoveNAs-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PipeOpRemoveNAs-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipeOpRemoveNAs$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
