% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{PipeOpRemoveZeros}
\alias{PipeOpRemoveZeros}
\title{PipeOpRemoveZeros}
\description{
Removes features from a Task for which the proportion
of zero values (\code{0}) are above a certain \code{cutoff} threshold.
Useful when the task has read count expression data (e.g. mRNA data).
}
\section{Parameters}{

\itemize{
\item \code{cutoff} :: \code{numeric(1)}\cr
Features with more than \code{cutoff} percentage of zeros will be removed.
Default value: 0.2.
}
}

\section{Initialization}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{PipeOpRemoveZeros$new()
porz = po('removezeros')
}\if{html}{\out{</div>}}
}

\examples{
library(mlr3proba)
library(mlr3pipelines)

df = data.frame(
  time = c(1,2,3,4),
  status = c(0,1,0,1),
  X1 = c(999,0,0,0),
  X2 = c(23,NA,0,0)
)

task = as_task_surv(x = df, id = 'test', time = 'time', event = 'status')

porz = PipeOpRemoveZeros$new(param_vals = list(cutoff = 0.7))
porz$train(list(task))[[1L]] # X1 removed, X2 remained

# another way to do the same
porz = po('removezeros', cutoff = 0.7)
porz$train(list(task))[[1L]] # X1 removed, X2 remained

}
\section{Super classes}{
\code{\link[mlr3pipelines:PipeOp]{mlr3pipelines::PipeOp}} -> \code{\link[mlr3pipelines:PipeOpTaskPreproc]{mlr3pipelines::PipeOpTaskPreproc}} -> \code{PipeOpRemoveZeros}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PipeOpRemoveZeros-new}{\code{PipeOpRemoveZeros$new()}}
\item \href{#method-PipeOpRemoveZeros-ncolsZero}{\code{PipeOpRemoveZeros$ncolsZero()}}
\item \href{#method-PipeOpRemoveZeros-clone}{\code{PipeOpRemoveZeros$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="help"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-help'><code>mlr3pipelines::PipeOp$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="predict"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-predict'><code>mlr3pipelines::PipeOp$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="print"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-print'><code>mlr3pipelines::PipeOp$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="train"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-train'><code>mlr3pipelines::PipeOp$train()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipeOpRemoveZeros-new"></a>}}
\if{latex}{\out{\hypertarget{method-PipeOpRemoveZeros-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipeOpRemoveZeros$new(id = "remove_zeros", param_vals = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character(1)})\cr
Identifier of the resulting  object.}

\item{\code{param_vals}}{(\code{list()})\cr
List of hyperparameter settings, overwriting the hyperparameter settings that would
otherwise be set during construction.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipeOpRemoveZeros-ncolsZero"></a>}}
\if{latex}{\out{\hypertarget{method-PipeOpRemoveZeros-ncolsZero}{}}}
\subsection{Method \code{ncolsZero()}}{
How many features (columns) have zeros (i.e. at least one)?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipeOpRemoveZeros$ncolsZero(task)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{\link[mlr3:Task]{Task}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipeOpRemoveZeros-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PipeOpRemoveZeros-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipeOpRemoveZeros$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
