% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survmob-package.R
\docType{package}
\name{survmob-package}
\alias{survmob}
\alias{survmob-package}
\title{survmob: Survival Multi-Omics Benchmarking}
\description{
Several functions for benchmarking survival models using the mlr3 framework.
}
\author{
\strong{Maintainer}: John Zobolas \email{bblodfon@gmail.com} (\href{https://orcid.org/0000-0002-3609-8674}{ORCID})

}
\keyword{internal}
