% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{tskSub}
\alias{tskSub}
\title{Subset task class}
\description{
Subset (survival) tasks to the most stable features identified
by the ensemble feature approach (\link{eFS}).
For a summary table of the number of features that can be selected via
various methods for cutoffs, etc. see \code{feat_stats()} and for actually
subsetting the tasks, see \code{subset_tasks()}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{nfeats}}{(\code{numeric(1)})\cr
Top N features}

\item{\code{perc}}{(\code{numeric(1)})\cr
Top N\% (percentage cutoff)}

\item{\code{cutoff}}{(\code{numeric(1)})\cr
Consensus Selection Frequency cutoff}

\item{\code{max_nfeats}}{(\code{numeric(1)})\cr
Max #features}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FSTaskSubsettor-new}{\code{tskSub$new()}}
\item \href{#method-FSTaskSubsettor-summary_stats}{\code{tskSub$summary_stats()}}
\item \href{#method-FSTaskSubsettor-subset_tasks}{\code{tskSub$subset_tasks()}}
\item \href{#method-FSTaskSubsettor-clone}{\code{tskSub$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FSTaskSubsettor-new"></a>}}
\if{latex}{\out{\hypertarget{method-FSTaskSubsettor-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tskSub$new(nfeats = 10, perc = 0.05, cutoff = 0.6, max_nfeats = 15)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nfeats}}{Top N features}

\item{\code{perc}}{Top N\% (percentage cutoff)}

\item{\code{cutoff}}{Selection Frequency cutoff}

\item{\code{max_nfeats}}{Max #features}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FSTaskSubsettor-summary_stats"></a>}}
\if{latex}{\out{\hypertarget{method-FSTaskSubsettor-summary_stats}{}}}
\subsection{Method \code{summary_stats()}}{
Summary table with the number of features that can be
selected based on available methods. See \code{subset_tasks()} for more
details.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tskSub$summary_stats(eFSlist, tasks)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{eFSlist}}{list of \link{eFS} objects}

\item{\code{tasks}}{list of \link[mlr3proba:TaskSurv]{TaskSurv} (1-1 correspondence with
each eFS object)}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Main sanity checks:
\itemize{
\item There has to be 1-1 correspondence between the tasks and the \link{eFS}
classes
\item The tasks have to be different (different id at least :)
\item The \link{eFS} objects must have executed the method \verb{$run()} so that there
is a \code{result} tibble
\item No task should have more than \code{max_nfeats} (see \code{initialize()}) features
}
}

\subsection{Returns}{
A \link[dplyr:reexports]{tibble}, with columns:
\itemize{
\item \code{task_id} => different task id per row
\item \code{features} => features names to select from (different vector per row,
ordered by consensus selection frequency)
\item \code{top_n} => #features to select based on \code{top_n} method
\item \code{top_nperc} => #features to select based on \code{top_nperc} method
\item \code{gt_freq} => #features to select based on \code{gt_freq} method
\item \code{score_median} => median performance score out of all feature subsets
used in eFS
\item \code{nfeats_median} => median #features selected out of all feature subsets
used in eFS
\item In case of a C-index measure, a combination measure and its normalized
version are included (\code{comb_score}, \code{comb_score_norm}) as well as a
column named \code{optim_nfeats}, see \code{subset_tasks()} for more info.
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FSTaskSubsettor-subset_tasks"></a>}}
\if{latex}{\out{\hypertarget{method-FSTaskSubsettor-subset_tasks}{}}}
\subsection{Method \code{subset_tasks()}}{
Subset tasks (i.e. reduce the number of features) based on
the corresponding consensus feature selection results (eFS)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tskSub$subset_tasks(eFSlist, tasks, method = "top_n")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{eFSlist}}{list of \link{eFS} objects}

\item{\code{tasks}}{list of \link[mlr3proba:TaskSurv]{TaskSurv} (1-1 correspondence with
each eFS object)}

\item{\code{method}}{method that decides which features to select from each task.
Available options are \code{top_n}, \code{top_nperc}, \code{gt_freq} or \code{optim_nfeats}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
To decide which features to select for a particular task, we use
the consensus frequency selection statistics from the corresponding \link{eFS}
object (features are ranked from most frequently selected to least).
This means that at least two RSFs were used during the execution of
the eFS algorithm.

Available options for \code{method}:
\itemize{
\item \code{top_n} => Select the top N (\code{nfeats}) features.
After subsetting, all tasks will have the same number of features.
\item \code{top_nperc} => Select the top N\% (\code{perc}) features.
This depends on the total number of features selected from each task by
the respective eFS, so the number of the top 5\% features for one task
might be different than the number of the top 5\% features in another task.
\item \code{gt_freq} => Select features whose selection frequency in the eFS
was above X\% (X is the \code{cutoff}).
\item \code{optim_nfeats} => The number of features for each task is decided based
on the median performance and median number of selected features in the
eFS (\strong{the larger} the both of these are, the more features will be
assigned to a task).
This method can be used only when all the eFS were executed with an
optimization measure that's a \strong{C-index} or \strong{OOB = 1 - C-index}.
So, one task will get a normalized \code{comb_score_norm} equal to 1 (larger
product of C-index * #features) and this task will get \code{max_nfeats} (from
class initialization) - the rest of the tasks will get less #features.
}
}

\subsection{Returns}{
list of \link[mlr3proba:TaskSurv]{TaskSurv}s with only the most
frequently selected features based on the consensus eFS results
(how many features is decided by the \code{method} and the result of
\code{summary_stats()})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FSTaskSubsettor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FSTaskSubsettor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tskSub$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
