% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{bench_msrs}
\alias{bench_msrs}
\title{Measures for Benchmarking Survival Models}
\usage{
bench_msrs()
}
\value{
a \link[data.table:data.table]{data.table} with the suggested survival
measures used for benchmarking purposes
}
\description{
A convenient function that returns a table with measures that
can be used for benchmarking survival learners.
}
\details{
The following measures are provided via \link{mlr3proba}:
\itemize{
\item Discrimination measures (ranking \code{crank} prediction type)
\itemize{
\item \link[mlr3proba:mlr_measures_surv.cindex]{Harrell's C-index} (doesn't account for censoring)
\item \link[mlr3proba:mlr_measures_surv.cindex]{Uno's C-index} (estimates censoring distribution using KM)
}
\item Calibration measures (survival \code{distr} prediction type)
\itemize{
\item \link[mlr3proba:mlr_measures_surv.graf]{Integrated Brier Score} (with \emph{proper = TRUE, method = 2}, estimates KM censoring distribution, is also a measure of discrimination)
\item \link[mlr3proba:mlr_measures_surv.rcll]{Right-Censored Log-Loss} (proper calibration score)
\item \link[mlr3proba:mlr_measures_surv.dcalib]{D-Calibration} (with \emph{B = 10, chisq = FALSE}, returning the measure and not the p-value)
}
}

Any of the above measures \strong{can} be used for model \emph{evaluation} (but depends
on several other factors if they \strong{should} be used, e.g. censoring
distribution, dataset size, proportion of events in the test set, etc.).
For model \emph{optimization/tuning}, we suggest the Right-Censored Log-Loss
for survival predictions and Uno's C-index for risk predictions
(e.g. linear predictors).
}
\examples{
library(mlr3proba)
ms = bench_msrs()
ms[id \%in\% c('ibrier','harrell_c')]$measure

}
