% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootstrapResult.R
\name{BootRes}
\alias{BootRes}
\title{Bootsrap Result Class}
\description{
mpla mpla
}
\examples{
mpla mpla
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{task_id}}{mpla}

\item{\code{lrn_id}}{mpla}

\item{\code{test_measure_ids}}{mpla}

\item{\code{test_nrsmps}}{mpla}

\item{\code{test_workers}}{mpla}

\item{\code{score}}{mpla}

\item{\code{scores}}{mpla}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BootstrapResult-new}{\code{BootRes$new()}}
\item \href{#method-BootstrapResult-calculate}{\code{BootRes$calculate()}}
\item \href{#method-BootstrapResult-percent_ci}{\code{BootRes$percent_ci()}}
\item \href{#method-BootstrapResult-score_median}{\code{BootRes$score_median()}}
\item \href{#method-BootstrapResult-clone}{\code{BootRes$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BootstrapResult-new"></a>}}
\if{latex}{\out{\hypertarget{method-BootstrapResult-new}{}}}
\subsection{Method \code{new()}}{
mpla mpla
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BootRes$new(
  test_measure_ids = bench_msrs()$id,
  test_nrsmps = 1000,
  test_workers = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{test_measure_ids}}{mpla}

\item{\code{test_nrsmps}}{mpla}

\item{\code{test_workers}}{mpla}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BootstrapResult-calculate"></a>}}
\if{latex}{\out{\hypertarget{method-BootstrapResult-calculate}{}}}
\subsection{Method \code{calculate()}}{
Calculate bootstrapping performance scores
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BootRes$calculate(task, learner, part, quiet = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{mpla}

\item{\code{learner}}{mpla}

\item{\code{part}}{mpla}

\item{\code{quiet}}{Default: TRUE => don't report time elapsed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BootstrapResult-percent_ci"></a>}}
\if{latex}{\out{\hypertarget{method-BootstrapResult-percent_ci}{}}}
\subsection{Method \code{percent_ci()}}{
Percentile confidence intervals per measure
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BootRes$percent_ci(conf = 0.95)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{conf}}{Confidence level. Default: 0.95 (i.e. the 2.5th and 97.5th
percentile points are used to define the confidence interval range)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BootstrapResult-score_median"></a>}}
\if{latex}{\out{\hypertarget{method-BootstrapResult-score_median}{}}}
\subsection{Method \code{score_median()}}{
Median summary score for the bootstrapping scores
(one per measure)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BootRes$score_median()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BootstrapResult-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BootstrapResult-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BootRes$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
