% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{minimize_backend}
\alias{minimize_backend}
\title{Minimize Backend of Survival Task}
\usage{
minimize_backend(task)
}
\arguments{
\item{task}{\link[mlr3proba:TaskSurv]{TaskSurv}}
}
\description{
Use this function when the backend of a task hasn't changed
after a lot of preprocessing, making it too heavy to use in further
modeling steps
}
