% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{reshape_mob_res}
\alias{reshape_mob_res}
\title{Reshape MOBenchmark result object to tibble}
\usage{
reshape_mob_res(res, add_modality_columns = TRUE, endfix = "_omic")
}
\arguments{
\item{res}{a \code{tibble} with columns \code{task_id}, \code{lrn_id} and \code{boot_res}.
Every other column is discarded.}

\item{add_modality_columns}{whether to generate individual columns for each
modality listed in the \code{task_id} column.
Modality names should be separated by a dash ('-') in multi-modal \code{task_id}s
and have different names.
The modality columns will have values 0 or 1, indicating the presence or
absence of a corresponding modality in the \code{task_id} column.
Default: \code{TRUE}.}

\item{endfix}{string to add to the end of each newly generated modality
column.
Default: \verb{_omic}.}
}
\value{
\code{tibble} with columns \code{task_id}, \code{lrn_id}, \code{rsmp_id}, \code{measure},
\code{value} and possibly one column per omic/modality with values 1 or 0.
}
\description{
Reshape the output \code{result} object from \link{MOBenchmark} (see example there).
}
