% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{task_powerset}
\alias{task_powerset}
\title{Create task powerset}
\usage{
task_powerset(tasks)
}
\arguments{
\item{`tasks`}{list of \link[mlr3:Task]{mlr3::Task}s}
}
\description{
Given a list of \link[mlr3:Task]{tasks}, this function produces the
set (list) of all possible combinations of tasks (powerset), combining their
respective features via \link[mlr3pipelines:mlr_pipeops_featureunion]{mlr3pipelines::mlr_pipeops_featureunion}.
Every task in the given list must have:
\itemize{
\item \strong{Different feature names and task ids}
\item Same target names (target columns only from the first task are kept)
\item Same number of observations/rows
}
}
